refactor: SelectedBoundsContext
