+++ packages/frontend/src/contexts/CanvasViewContext.tsx
@@ -7,12 +7,12 @@ import {
   RefObject,
   SetStateAction,
   useContext,
+  useEffect,
   useMemo,
   useRef,
   useState,
 } from "react";
 import { addPoints, ORIGIN, tupleToPoint } from "@/components/canvas/point";
-import { ViewBounds } from "@/util";
 import { useCanvasContext } from "./CanvasContext";
 
 interface CanvasViewContextType {
@@ -21,12 +21,10 @@ interface CanvasViewContextType {
   coords: Point | null;
   isReticleVisible: boolean;
   offset: Point;
-  selectedBounds: ViewBounds | null;
   zoom: number;
   setCoords: Dispatch<SetStateAction<Point | null>>;
   setIsReticleVisible: Dispatch<SetStateAction<boolean>>;
   setOffset: Dispatch<SetStateAction<Point>>;
-  setSelectedBounds: Dispatch<SetStateAction<ViewBounds | null>>;
   setZoom: Dispatch<SetStateAction<number>>;
 }
 
@@ -36,12 +34,10 @@ export const CanvasViewContext = createContext<CanvasViewContextType>({
   coords: null,
   isReticleVisible: false,
   offset: ORIGIN,
-  selectedBounds: null,
   zoom: 1,
   setCoords: () => {},
   setIsReticleVisible: () => {},
   setOffset: () => {},
-  setSelectedBounds: () => {},
   setZoom: () => {},
 });
 
@@ -58,8 +54,6 @@ export const CanvasViewProvider = ({ children }: CanvasViewProviderProps) => {
   const [zoom, setZoom] = useState(1);
   const containerRef = useRef<HTMLDivElement | null>(null);
   const [offset, setOffset] = useState(ORIGIN);
-  const [selectedBounds, setSelectedBounds] =
-    useState<CanvasViewContextType["selectedBounds"]>(null);
 
   const adjustedCoords = useMemo(() => {
     if (selectedCoords) {
@@ -69,6 +63,14 @@ export const CanvasViewProvider = ({ children }: CanvasViewProviderProps) => {
     return null;
   }, [canvas.startCoordinates, selectedCoords]);
 
+  // biome-ignore lint/correctness/useExhaustiveDependencies: Trigger on canvas change
+  useEffect(
+    function resetCoordsOnCanvasChange() {
+      setSelectedCoords(null);
+    },
+    [canvas.id],
+  );
+
   return (
     <CanvasViewContext.Provider
       value={{
@@ -77,12 +79,10 @@ export const CanvasViewProvider = ({ children }: CanvasViewProviderProps) => {
         coords: selectedCoords,
         isReticleVisible: isReticleVisible && selectedCoords !== null,
         offset: offset,
-        selectedBounds,
         zoom: zoom,
         setCoords: setSelectedCoords,
         setIsReticleVisible: setIsReticleVisible,
         setOffset: setOffset,
-        setSelectedBounds: setSelectedBounds,
         setZoom: setZoom,
       }}
     >
