+++ packages/frontend/src/contexts/index.ts
@@ -3,6 +3,10 @@ export { CanvasProvider, useCanvasContext } from "./CanvasContext";
 export { CanvasViewProvider, useCanvasViewContext } from "./CanvasViewContext";
 export { QueryClientProvider } from "./QueryClientProvider";
 export {
+  SelectedBoundsProvider,
+  useSelectedBoundsContext,
+} from "./SelectedBoundsContext";
+export {
   SelectedColorProvider,
   useSelectedColorContext,
 } from "./SelectedColorContext";
